refactor: move canonical number calc to block-processor
