+++ packages/core/src/node/zkopru-node.ts
@@ -3,7 +3,6 @@ import { ZkAccount } from '@zkopru/account'
 import { DB } from '@zkopru/database'
 import { Grove, poseidonHasher, keccakHasher } from '@zkopru/tree'
 import { logger } from '@zkopru/utils'
-import AsyncLock from 'async-lock'
 import { L1Contract } from '../context/layer1'
 import { L2Chain } from '../context/layer2'
 import { BootstrapHelper } from './bootstrap'
@@ -11,7 +10,6 @@ import { Synchronizer } from './synchronizer'
 import { Tracker } from './tracker'
 import { BlockProcessor } from './block-processor'
 import { Watchdog } from './watchdog'
-import { Block } from '../block'
 
 export class ZkopruNode {
   running: boolean
@@ -36,8 +34,6 @@ export class ZkopruNode {
   watchdog?: Watchdog
 
   bootstrapHelper?: BootstrapHelper
-
-  canonicalLock = new AsyncLock()
 
   constructor({
     db,
@@ -89,10 +85,6 @@ export class ZkopruNode {
       })
       this.blockProcessor.on('processed', async proposal => {
         this.synchronizer.setLatestProcessed(proposal.proposalNum)
-        await this.calcCanonicalBlockHeights()
-      })
-      this.synchronizer.on('status', async () => {
-        await this.calcCanonicalBlockHeights()
       })
     } else {
       logger.info('already on syncing')
@@ -153,74 +145,5 @@ export class ZkopruNode {
     })
     await grove.init()
     return new L2Chain(db, grove, config, vks)
-  }
-
-  // idempotently calculate canonical numbers
-  private async calcCanonicalBlockHeights() {
-    await this.canonicalLock.acquire('canon', async () => {
-      // find earliest block with no canonical num
-      const startBlock = await this.db.findMany('Proposal', {
-        where: {
-          canonicalNum: null,
-          OR: [{ proposalData: { ne: null } }, { proposalNum: 0 }],
-        },
-        orderBy: { proposalNum: 'asc' },
-        limit: 1,
-      })
-      if (startBlock.length === 0) {
-        // have canonical numbers for all blocks
-        return
-      }
-      // The proposal to start at
-      const [{ proposalNum }] = startBlock
-      if (proposalNum === null) {
-        throw new Error('Proposal number is null')
-      }
-      const blockHeight = await this.db.count('Proposal', {})
-      const latestProcessed = this.synchronizer.latestProcessed || 0
-      for (
-        let x = proposalNum;
-        x <= Math.min(blockHeight, latestProcessed);
-        x += 1
-      ) {
-        await this.calcCanonicalBlockHeight(x)
-      }
-    })
-  }
-
-  private async calcCanonicalBlockHeight(proposalNum: number) {
-    const proposals = await this.db.findMany('Proposal', {
-      where: { proposalNum },
-    })
-    if (proposals.length !== 1) {
-      throw new Error(`Did not find one proposal for number: ${proposalNum}`)
-    }
-    const [proposal] = proposals
-    const { hash } = proposal
-    if (proposalNum === 0) {
-      await this.db.update('Proposal', {
-        where: { hash },
-        update: { canonicalNum: 0 },
-      })
-      return
-    }
-    if (!proposal.proposalData) return
-    const block = Block.fromTx(JSON.parse(proposal.proposalData))
-    const header = block.getHeaderSql()
-    const parent = await this.db.findOne('Proposal', {
-      where: {
-        hash: header.parentBlock.toString(),
-      },
-    })
-    if (!parent) {
-      throw new Error(`Unable to find parent proposal`)
-    }
-    if (parent.canonicalNum === null) {
-      throw new Error(`Expected canonicalNum to exist!`)
-    }
-    await this.db.update('Proposal', {
-      where: { hash },
-      update: { canonicalNum: (parent.canonicalNum as number) + 1 },
-    })
   }
 }
